% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_finder.R
\name{cont_finder}
\alias{cont_finder}
\title{A continent finder.}
\usage{
cont_finder(country_wanted)
}
\arguments{
\item{country}{Write down any country you would like! (as long as it's in the database that is)}
}
\value{
The continent your country resides in.
}
\description{
This function allows you to find out what continent your country of choice resides in.
}
\examples{
cont.finder("x")
}
