% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_finder.R
\name{cou_finder}
\alias{cou_finder}
\title{A country finder.}
\usage{
cou_finder(continent_wanted)
}
\arguments{
\item{continent}{Write down any continent you would like!}
}
\value{
The countries that are part of your continent.
}
\description{
This function allows you to find out what countries are part of the given continent.
}
\examples{
cou.finder("x")
}
