% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_expectancy_graph.R
\name{life_expectancy_graph}
\alias{life_expectancy_graph}
\title{A graph that shows the life expectancy in a specific year, in a specific continent.}
\usage{
life_expectancy_graph(continent_wanted, year_wanted)
}
\arguments{
\item{continent, year}{Make sure to write down both the year AND continent wanted.The year can only be from 2002-2015, since the database doesn't go any further than those years.}
}
\value{
A bar graph showing the difference in life expectancy between the countries.
}
\description{
This function allows you to find out what the difference is between the life expectancy, taken from countries all residing in the same continent, in the same year.
}
\examples{
life_expectancy_graph("x","y")
}
