% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_finder.R
\name{population_finder}
\alias{population_finder}
\title{A total population finder.}
\usage{
population_finder(country, year)
}
\arguments{
\item{region}{Write down any country and year between 2002 and 2015 you would like! (as long as it's in the database that is)}
}
\value{
The total population during the requested year of your requested country.
}
\description{
This function allows you to find the total population for the country, and during the year of your choice.
}
\examples{
Population_finder("x","y")
}
